+++ packages/create-app-shared/CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## 0.2.4
 
-- [fix] ts types of runtime module
 - [feat] add runtime api of wrapperRouterRender
 - [fix] remove `deepmerge`, and fix deep clone let custom config `__proto__` lost
+- [fix] ts types of runtime module
 
 ## 0.2.3