+++ packages/create-app-shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.4
+
+- [feat] add runtime api of wrapperRouterRender
+- [fix] remove `deepmerge`, and fix deep clone let custom config `__proto__` lost
+- [fix] ts types of runtime module
+
 ## 0.2.3
 
 - [fix] ts types of export apis